+++ cmake/nfxSiliconSources.cmake
@@ -4,6 +4,10 @@
 
 set(NFX_SILICON_SOURCES
     # Discrete components
+    src/discrete/Capacitor.cpp
+    src/discrete/Inductor.cpp
+    src/discrete/Resistor.cpp
+
     src/discrete/PowerSupply.cpp
 
     # Signal layer